ant colony integration

+++ antcolonythread.h
@@ -11,14 +11,13 @@ class AntColonyThread : public QThread
     Q_OBJECT
 public:
     AntColonyThread();
-    AntColonyThread(QString fileName, int seed, int populationSize, int generations, double probability_q,
-                                     double q, int number_mods, double evaporation, double influence);
+    AntColonyThread(QString fileName, int seed, int populationSize, int generations, int probability_q,
+                                     int q, int number_mods, int evaporation, int influence);
     void run();
 private:
     Population population;
     Problem problem;
     AntColonyAlgorithm algorithm;
-    //QElapsedTimer timer;
     QString fileName;
     int seed;
     int populationSize;
